Add files via upload
This project involved the implementation of fundamental machine learning algorithms, specifically Decision Trees and Random Forests, from scratch. The primary objective was to apply these algorithms to a dataset and evaluate their performance in both classification and regression tasks. By manually coding these models, the project provided a deep understanding of how these algorithms operate, including the intricacies of decision-making within trees and the power of ensemble methods like Random Forests. The project included data preprocessing, model training, and evaluation using various metrics, as well as visualizations to analyze the results. This hands-on approach reinforced core machine learning concepts and demonstrated the practical application of these algorithms in predictive modeling. Additionally, the presentation component of the project focused on highlighting a few surprising findings that emerged during the analysis, which were beyond the scope of the initial questions. These observations provided valuable insights and added depth to the overall project experience.